Moved model-based utils to common
