+++ morl_baselines/multi_policy/gpi_pd/gpi_pd_continuous_action.py
@@ -13,6 +13,10 @@ import wandb as wb
 from mo_gymnasium.evaluation import policy_evaluation_mo
 
 from morl_baselines.common.buffer import ReplayBuffer
+from morl_baselines.common.model_based.probabilistic_ensemble import (
+    ProbabilisticEnsemble,
+)
+from morl_baselines.common.model_based.utils import ModelEnv, visualize_eval
 from morl_baselines.common.morl_algorithm import MOAgent, MOPolicy
 from morl_baselines.common.networks import mlp
 from morl_baselines.common.performance_indicators import expected_utility
@@ -23,10 +27,6 @@ from morl_baselines.common.utils import (
     log_episode_info,
     polyak_update,
 )
-from morl_baselines.multi_policy.gpi_pd.model.probabilistic_ensemble import (
-    ProbabilisticEnsemble,
-)
-from morl_baselines.multi_policy.gpi_pd.model.utils import ModelEnv, visualize_eval
 from morl_baselines.multi_policy.linear_support.linear_support import LinearSupport
