--- morl_baselines/multi_policy/gpi_pd/model/__init__.py
@@ -1,1 +1,0 @@
-"""Model for GPI-PD."""
